requirements.txt
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dash==2.0.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.1
Flask-Compress==1.10.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.21.2
pandas==1.3.3
plotly==5.3.1
pylotoncycle==0.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
sourcery-cli==0.10.1
tenacity==8.0.1
tomli==2.0.0
typing-extensions==4.0.1
urllib3==1.26.7
waiting==1.4.1
Werkzeug==2.0.1
